production_ready/requirements.txt
# Skin Analyzer API - Production Requirements
# Optimized for Google Cloud Run deployment

# Core Flask dependencies
Flask==3.0.0
Werkzeug==3.0.1
flask-cors==4.0.0

# Machine Learning
scikit-learn==1.5.2
joblib==1.4.2
numpy==1.26.4

# Image processing
opencv-python-headless==4.8.1.78
Pillow==10.1.0

# HTTP client (for testing)
requests==2.31.0

# Production server
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0